Rework API to use js exceptions for errors & generalize conversions.
This is a somewhat significant rework of the API to improve the
usability of address parsing. Specifically, `get_ua_receivers`
has been modified so that it can be used to parse arbitrary Zcash
addresses.

A common use case is to attempt to extract the transparent component of
a unified or TEX address, so that a bitcoin-derived transparent-only
wallet can send to the address. This is now facilitated via the
`getAddressReceivers` method. It returns a data structure with a
nullable field for each possible receiver type, with non-null fields
encoded as valid Zcash addresses of the most primitive possible type.